+++ SumCalculatorTests/CalcNoteTests.swift
@@ -27,13 +27,13 @@ class CalcNoteTests: XCTestCase {
         let calcItem1 = CalcItem()
         calcItem1.quantity = 3
         calcItem1.unitPrice = 1000
-        let testResult1:Float = 3000
+        let testResult1:Double = 3000
         calcTable1.calcItems.append(calcItem1)
         
         let calcItem2 = CalcItem()
         calcItem2.quantity = 4
         calcItem2.unitPrice = 2000
-        let testResult2:Float = 8000
+        let testResult2:Double = 8000
         calcTable1.calcItems.append(calcItem2)
         
         calcNote.calcTables.append(calcTable1)
@@ -43,7 +43,7 @@ class CalcNoteTests: XCTestCase {
         let calcItem3 = CalcItem()
         calcItem3.quantity = 2
         calcItem3.unitPrice = 4000
-        let testResult3:Float = 8000
+        let testResult3:Double = 8000
         calcTable2.calcItems.append(calcItem3)
         
         calcNote.calcTables.append(calcTable2)
@@ -53,7 +53,7 @@ class CalcNoteTests: XCTestCase {
         let calcItem4 = CalcItem()
         calcItem4.quantity = 1
         calcItem4.unitPrice = 5000
-        let testResult4:Float = 5001
+        let testResult4:Double = 5001
         calcTable2.calcItems.append(calcItem4)
         
         XCTAssertNotEqual(calcNote.total, testResult1 + testResult2 + testResult3 + testResult4)
